packages/voice-daemon/requirements.txt
# Voice Daemon
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
numpy>=1.24.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
vosk>=0.3.38
faster-whisper>=0.10.0
httpx>=0.25.0
python-dotenv>=1.0.0
aiofiles>=23.2.0

# ── Upgraded STT/TTS ──
edge-tts>=7.0.0
torch>=2.0.0
torchaudio>=2.0.0
